# backend/src/main/resources/openapi/switchboard-api.yaml
openapi: 3.0.3
info:
  title: Switchboard API
  description: AI-powered feature flag platform. Management surface (user bearer auth), SDK surface (SDK-key auth), and job triggers.
  version: 0.1.0
servers:
  - url: http://localhost:28080
security:
  - bearerAuth: []
...
paths:
...
  /api/projects/{projectId}/environments:
    get:
      tags: [Environments]
      operationId: listEnvironments
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Environments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentResponse'
    post:
      tags: [Environments]
      operationId: createEnvironment
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '201':
          description: Created environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
  /api/environments/{envId}/sdk-keys:
    post:
      tags: [Environments]
      operationId: createSdkKey
      parameters:
        - $ref: '#/components/parameters/EnvId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SdkKeyCreateRequest'
      responses:
        '201':
          description: Created key. The full key is returned ONLY here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkKeyCreatedResponse'
    get:
      tags: [Environments]
      operationId: listSdkKeys
      parameters:
        - $ref: '#/components/parameters/EnvId'
      responses:
        '200':
          description: Keys (prefixes only)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SdkKeyResponse'
  /api/sdk-keys/{keyId}:
    delete:
      tags: [Environments]
      operationId: revokeSdkKey
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
...
  /api/environments/{envId}/approval-settings:
    get:
      tags: [ChangeRequests]
      operationId: getApprovalSettings
      summary: This environment's approval policy.
      parameters:
        - $ref: '#/components/parameters/EnvId'
      responses:
        '200':
          description: Approval policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalSettingsResponse'
    put:
      tags: [ChangeRequests]
      operationId: updateApprovalSettings
      summary: Turn approval on or off for this environment. Requires MANAGE_ENVIRONMENTS.
      description: >
        Every field defaults to off, so an environment nobody has configured
        behaves exactly as it did before change requests existed. Omitted fields
        are left unchanged.
      parameters:
        - $ref: '#/components/parameters/EnvId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalSettingsUpdateRequest'
      responses:
        '200':
          description: The stored policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalSettingsResponse'
  /api/projects/{projectId}/change-requests:
    get:
      tags: [ChangeRequests]
      operationId: listChangeRequests
      summary: Change requests in this project, newest first (keyset cursor).
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: envKey
          in: query
          schema:
            type: string
        - name: flagKey
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ChangeRequestStatus'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Change requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestListResponse'
  /api/change-requests/{changeRequestId}:
    get:
      tags: [ChangeRequests]
      operationId: getChangeRequest
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
      responses:
        '200':
          description: One change request with its reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
  /api/change-requests/{changeRequestId}/approve:
    post:
      tags: [ChangeRequests]
      operationId: approveChangeRequest
      summary: Approve. Requires APPROVE_CHANGES in the request's environment.
      description: >
        The approval that meets minApprovals applies the change in the same call,
        through the normal versioned write path. If the flag has moved past the
        request's baseVersion the request goes STALE instead of overwriting the
        newer work, and the author has to rebase. Self-approval is refused unless
        the environment allows it.
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestDecisionRequest'
      responses:
        '200':
          description: The change request after the review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '403':
          description: No APPROVE_CHANGES here, or self-approval is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request is no longer PENDING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/change-requests/{changeRequestId}/decline:
    post:
      tags: [ChangeRequests]
      operationId: declineChangeRequest
      summary: Decline. One decline settles it. Requires APPROVE_CHANGES.
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestDecisionRequest'
      responses:
        '200':
          description: The declined change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
  /api/change-requests/{changeRequestId}/withdraw:
    post:
      tags: [ChangeRequests]
      operationId: withdrawChangeRequest
      summary: Withdraw your own pending request.
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
      responses:
        '200':
          description: The withdrawn change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '403':
          description: Only the author may withdraw
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/change-requests/{changeRequestId}/apply:
    post:
      tags: [ChangeRequests]
      operationId: applyChangeRequest
      summary: Idempotent retry for an APPROVED request whose write did not land.
      description: >
        Approvals auto-apply when the threshold is met, so this exists only for the
        case where that write failed. It refuses anything not already APPROVED, so
        it is not a way around review.
      parameters:
        - $ref: '#/components/parameters/ChangeRequestId'
      responses:
        '200':
          description: The applied (or now STALE) change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '409':
          description: The request is not APPROVED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Firebase ID token, or dev token (Bearer dev:email) in local profile.
    sdkKeyAuth:
      type: http
      scheme: bearer
      description: Environment SDK key (sb_srv_...).
    ofrepApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        OFREP's apiKey scheme, carrying the same environment SDK key (sb_srv_...) as sdkKeyAuth.
        Both headers resolve to the same SdkKeyPrincipal; Authorization wins when both are sent.
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EnvId:
      name: envId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChangeRequestId:
      name: changeRequestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum: [NOT_FOUND, FORBIDDEN, CONFLICT, VALIDATION_FAILED, UNAUTHORIZED, AI_UNAVAILABLE, UPGRADE_REQUIRED]
        message:
          type: string
...
    EnvironmentCreateRequest:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          pattern: '^[a-z][a-z0-9-]*$'
          maxLength: 64
        name:
          type: string
    EnvironmentResponse:
      type: object
      required: [id, projectId, key, name, stateVersion]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        stateVersion:
          type: integer
          format: int64
        approvals:
          $ref: '#/components/schemas/ApprovalSettingsResponse'
    SdkKeyCreateRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 100
    SdkKeyCreatedResponse:
      type: object
      required: [id, environmentId, key, keyPrefix, createdAt]
      properties:
        id:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        key:
          type: string
          description: Full SDK key. Returned only at creation.
        keyPrefix:
          type: string
        label:
          type: string
        createdAt:
          type: string
          format: date-time
    SdkKeyResponse:
      type: object
      required: [id, environmentId, keyPrefix, createdAt]
      properties:
        id:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        keyPrefix:
          type: string
        label:
          type: string
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
...
    ClauseOp:
      type: string
      enum: [EQUALS, IN, CONTAINS, STARTS_WITH, SEGMENT_MATCH, NOT_SEGMENT_MATCH]
    Clause:
      type: object
      required: [attribute, op, values]
      properties:
        attribute:
          type: string
          description: Context attribute name; 'key' targets the context key. For SEGMENT_MATCH ops, values carry segment keys.
        op:
          $ref: '#/components/schemas/ClauseOp'
        values:
          type: array
          items:
            type: string
    WeightedVariation:
      type: object
      required: [variationId, weight]
      properties:
        variationId:
          type: string
          format: uuid
        weight:
          type: integer
          minimum: 0
          maximum: 100
    RolloutOrVariation:
      type: object
      description: Exactly one of variationId or rollout must be set.
      properties:
        variationId:
          type: string
          format: uuid
        rollout:
          type: array
          description: Weights must sum to exactly 100.
          items:
            $ref: '#/components/schemas/WeightedVariation'
    IndividualTarget:
      type: object
      required: [contextKey, variationId]
      properties:
        contextKey:
          type: string
        variationId:
          type: string
          format: uuid
    Rule:
      type: object
      required: [id, clauses, serve]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        clauses:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Clause'
        serve:
          $ref: '#/components/schemas/RolloutOrVariation'
    FlagTargetingConfig:
      type: object
      required: [offVariationId, defaultVariationId, fallthrough]
      properties:
        individualTargets:
          type: array
          items:
            $ref: '#/components/schemas/IndividualTarget'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
        fallthrough:
          $ref: '#/components/schemas/RolloutOrVariation'
        offVariationId:
          type: string
          format: uuid
        defaultVariationId:
          type: string
          format: uuid
...
    ApprovalSettingsResponse:
      type: object
      required: [requireApproval, minApprovals, allowSelfApproval, requireApprovalForKill, allowAutomationBypass]
      properties:
        requireApproval:
          type: boolean
          description: Targeting writes and rollbacks open a change request instead of writing.
        minApprovals:
          type: integer
          description: How many reviewers other than the author (unless allowSelfApproval) must approve.
        allowSelfApproval:
          type: boolean
          description: When false the author's own approval is refused outright.
        requireApprovalForKill:
          type: boolean
          description: >
            Opt the kill switch into review as well. Off by default: the emergency
            stop bypasses approval so an incident cannot be blocked on a reviewer.
        allowAutomationBypass:
          type: boolean
          description: >
            ON BY DEFAULT, AND THIS IS THE ONE SETTING THAT LEAVES AN UNREVIEWED
            WRITE PATH OPEN. It only means anything when requireApproval is on, and
            only automation can use it: the rollout monitor's auto-rollback and
            auto-optimize keep writing immediately instead of opening a change
            request. A HUMAN applying an AI proposal is never covered - they are
            gated exactly like a hand edit. The default mirrors the kill switch:
            an automated rollback fires during an error spike and puts traffic back
            on the baseline variation that was already live, so making it wait for a
            reviewer removes the reason to automate it. Set it to false and automated
            healing parks in the review queue like everything else. Either way the
            write is audited, and a bypassed write additionally records an
            APPROVAL_BYPASS entry naming the automation as the actor.
    ApprovalSettingsUpdateRequest:
      type: object
      description: Omitted fields are left unchanged.
      properties:
        requireApproval:
          type: boolean
        minApprovals:
          type: integer
          minimum: 1
          maximum: 10
        allowSelfApproval:
          type: boolean
        requireApprovalForKill:
          type: boolean
        allowAutomationBypass:
          type: boolean
          description: >
            See ApprovalSettingsResponse.allowAutomationBypass. Defaults to true on
            every environment; turning it off is what makes automated rollout healing
            wait for review in a gated environment.
    ChangeRequestKind:
      type: string
      enum: [TARGETING_UPDATE, KILL_SWITCH, ROLLBACK]
    ChangeRequestStatus:
      type: string
      description: >
        PENDING is the only reviewable state. STALE means the flag moved past the
        request's baseVersion before it could be applied, so it was refused rather
        than allowed to clobber the newer version.
      enum: [PENDING, APPROVED, DECLINED, APPLIED, WITHDRAWN, STALE]
    ReviewDecision:
      type: string
      enum: [APPROVE, DECLINE]
    ChangeRequestPayload:
      type: object
      description: >
        The proposed write, in the shape the flag write endpoints take. Exactly the
        fields of one kind are populated - enabled+config for TARGETING_UPDATE,
        active for KILL_SWITCH, toVersion for ROLLBACK.
      properties:
        enabled:
          type: boolean
        config:
          $ref: '#/components/schemas/FlagTargetingConfig'
        active:
          type: boolean
        toVersion:
          type: integer
    ChangeRequestReviewResponse:
      type: object
      required: [id, reviewerUserId, reviewer, decision, createdAt]
      properties:
        id:
          type: string
          format: uuid
        reviewerUserId:
          type: string
          format: uuid
        reviewer:
          type: string
        decision:
          $ref: '#/components/schemas/ReviewDecision'
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChangeRequestResponse:
      type: object
      required: [id, orgId, projectId, environmentId, envKey, flagId, flagKey, kind, payload, baseVersion, minApprovals, allowSelfApproval, status, requestedBy, requestedByUserId, createdAt, reviews, approvalsMet]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        envKey:
          type: string
        flagId:
          type: string
          format: uuid
        flagKey:
          type: string
        kind:
          $ref: '#/components/schemas/ChangeRequestKind'
        payload:
          $ref: '#/components/schemas/ChangeRequestPayload'
        baseVersion:
          type: integer
          description: The head version the author edited against.
        minApprovals:
          type: integer
          description: Snapshotted from the environment when the request was opened.
        allowSelfApproval:
          type: boolean
        status:
          $ref: '#/components/schemas/ChangeRequestStatus'
        requestedByUserId:
          type: string
          format: uuid
        requestedBy:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
        appliedVersion:
          type: integer
          description: The config version this request produced once applied.
        aiProposalId:
          type: string
          format: uuid
          description: >
            Set when this request exists because an AI proposal was applied into an
            environment that requires approval. The proposal stays DRAFT until the
            request applies, and becomes APPLIED when the last of its requests lands.
        approvalsMet:
          type: integer
          description: Approvals that count toward the threshold right now.
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestReviewResponse'
    ChangeRequestListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestResponse'
        nextCursor:
          type: string
    ChangeRequestDecisionRequest:
      type: object
      properties:
        comment:
          type: string